Improve PATH setting code
Cleaner code that also works much more reliably and doesn't unnecessarily slow down deployments.

Instead of rendering the path we're going to prepend PATH with at deploy time, we do it at the time the user opens bash. This has the added benefit of PATH always being up to date, rather than referencing the path that was "latest and greatest" at deployment time.